+++ app.rb
@@ -5,67 +5,11 @@ require 'sinatra/base'
 require 'sprockets'
 require 'uglifier'
 
+require_relative 'lib/demo'
+require_relative 'lib/message'
+
 require_relative 'lib/loader'
-
 Loader.get :latest
-
-class Demo
-  attr_reader :rouge, :lexer, :source
-
-  def initialize(ver = :latest, lang = nil, source = nil)
-    @rouge = set_version ver
-    @lexer = set_lexer lang
-    @source = set_source source
-  end
-
-  def all_lexers
-    rouge::Lexer.all.sort_by(&:tag)
-  end
-
-  def lexer_count
-    all_lexers.count
-  end
-
-  def result
-    rouge.highlight source, lexer, 'html'
-  end
-
-  def version
-    rouge.version
-  end
-
-  private def set_version(ver)
-    ver = (ver.is_a?(String) && ver[0] == "v") ? ver.slice(1..-1) : :latest
-
-    Loader.get ver
-  end
-
-  private def set_lexer(lang)
-    return all_lexers.sample if lang.nil?
-
-    rouge::Lexer.find(lang) || all_lexers.sample
-  end
-
-  private def set_source(source)
-    source || lexer.demo
-  end
-end
-
-class Message
-  MSG = {}
-
-  MSG[400] =
-    "<strong>Bad Input</strong>: The input that you submitted is invalid. Please
-    correct it and try again."
-
-  MSG[413] =
-    "<strong>Too Long</strong>: This form accepts a maximum of 1500 characters of text.
-    Please reduce the amount of text and try again."
-
-  def self.[](k)
-    MSG[k]
-  end
-end
 
 class Dingus < Sinatra::Base
   enable :sessions
